Added Linked List
All boxes are now included in a linked list after they get instantiated.
The first box is instantiated outside of the while-game loop and added
to the linked list first.

Once a box is clicked or 'selected' it is immediately removed from the
linked list. This means I can bar it from being called on when I invoke
the box.move() function. Thereby preventing it from being moved
automatically with the user dragging it

Just have to be sure boxes that aren't clicked won't be affected when
other boxes are clicked and dragged / placed.