Overlap on Any Points in Rect
overlap() now detects if two rectangles intersect on any of their
pixels. Sometimes there seems to be glitches were a detection is not
triggered when two rects overlap but for the most part it's pretty
consistent. Now have to figure out how to prevent user from placing
a rect when overlap() is triggered. I can't return False from overlap()
because it will stop the iteration for each box in boxes. Unless I have
some if-statement that says if all boxes have been checked and none
intersect then return False